requirements.txt
gradio
opencv-python
Pillow
insightface
onnxruntime
numpy